+++ relm4-components/src/open_button/factory.rs
@@ -1,6 +1,6 @@
-use gtk::prelude::ButtonExt;
-use relm4::factory::{DynamicIndex, FactoryPrototype, FactoryVecDeque};
-use relm4::{gtk, send, WidgetPlus};
+use gtk::prelude::*;
+use relm4::factory::{DynamicIndex, FactoryComponent, FactoryComponentSender};
+use relm4::{gtk, WidgetPlus};
 
 use super::OpenButtonMsg;
 
@@ -11,53 +11,33 @@ pub(crate) struct FileListItem {
     pub(crate) path: PathBuf,
 }
 
-#[derive(Debug)]
-pub(crate) struct FileListItemWidgets {
-    label: gtk::Button,
-    row: gtk::ListBoxRow,
-}
-
-impl FactoryPrototype for FileListItem {
-    type Factory = FactoryVecDeque<Self>;
-    type View = gtk::Box;
+#[relm4::factory(pub(crate))]
+impl FactoryComponent for FileListItem {
+    type ParentInput = OpenButtonMsg;
+    type CommandOutput = ();
+    type Input = ();
+    type Init = PathBuf;
+    type ParentWidget = gtk::Box;
+    type Output = OpenButtonMsg;
     type Widgets = FileListItemWidgets;
-    type Root = gtk::ListBoxRow;
-    type Msg = OpenButtonMsg;
 
-    fn init_view(&self, key: &DynamicIndex, sender: relm4::Sender<Self::Msg>) -> Self::Widgets {
-        let label = gtk::Button::with_label(
-            self.path
-                .iter()
-                .last()
-                .expect("Empty path")
-                .to_str()
-                .expect("Couldn't convert path to string"),
-        );
-        let row = gtk::ListBoxRow::builder().child(&label).build();
-
-        label.inline_css(b"margin: 0");
-
-        let key = key.clone();
-        label.connect_clicked(move |_| {
-            send!(sender, OpenButtonMsg::OpenRecent(key.clone()));
-        });
-        FileListItemWidgets { label, row }
+    view! {
+        gtk::ListBoxRow {
+            gtk::Button {
+                set_label: self.path.iter().last().expect("Empty path").to_str().unwrap(),
+                set_margin_all: 0,
+                connect_clicked[sender, index] => move |_| {
+                    sender.output(OpenButtonMsg::OpenRecent(index.clone()));
+                }
+            }
+        }
     }
 
-    fn view(&self, _key: &DynamicIndex, widgets: &Self::Widgets) {
-        widgets.label.set_label(
-            self.path
-                .iter()
-                .last()
-                .expect("Empty path")
-                .to_str()
-                .expect("Couldn't convert path to string"),
-        );
+    fn output_to_parent_input(output: Self::Output) -> Option<Self::ParentInput> {
+        Some(output)
     }
 
-    fn position(&self, _key: &DynamicIndex) {}
-
-    fn root_widget(widgets: &Self::Widgets) -> &Self::Root {
-        &widgets.row
+    fn init_model(init: Self::Init, _: &DynamicIndex, _: FactoryComponentSender<Self>) -> Self {
+        Self { path: init }
     }
 }